refactoring to start use callofduty api
