+++ src/config.js
@@ -12,4 +12,14 @@ const api = axios.create({
   },
 });
 
-module.exports = { api };
+const myCallOfDuty = axios.create({
+  baseURL: 'https://my.callofduty.com',
+  timeout: 25000,
+});
+
+const profileApi = axios.create({
+  baseURL: 'https://profile.callofduty.com',
+  timeout: 25000,
+});
+
+module.exports = { api, myCallOfDuty, profileApi };
